docker/requirements.txt
# Core dependencies
fastapi>=0.104.0
uvicorn>=0.24.0
pydantic>=2.0.0
python-dotenv>=1.0.0

# AI clients
openai>=1.0.0

# Storage (R2/S3)
boto3>=1.34.0

# Repository management
GitPython>=3.1.40

# RepoMapper dependencies
tiktoken>=0.5.0
networkx>=3.0
diskcache>=5.6.0
grep-ast>=0.3.0
tree-sitter>=0.20.0
pygments>=2.14.0
fastmcp

# Other dependencies
langchain-core>=0.1.0
